Parse Checked C pointer types.
Add parsing of Checked C pointer types.  Map them to the existing pointer
type during semantic analysis.  These types have the form ptr<T> and
array_ptr<T>, where T is a type.  The code for parsing is modeled after
the code for parsing atomic type specifiers (which uses parentheses instead
of < >).  We have to be careful to handle the case of ptr<ptr<T>>, where
the >> is tokenized as a right shift operator.  ptr and array_ptr are treated
as keywords, simplifying parsing.

Also add some other new Checked C keywords.  The new keywords include where,
checked, and unchecked.

For now, we are living with the fact that these new keywords  may conflict
with existing identifier names.  The names ptr and checked are
particularly likely to have conflicts with existing variable names.
We'll sort that out in the future some day, if there are strong objections
from the community.  For now, we'll stick with keywords that are short and
visually easy to read.

A test and test harness support will be checked in separately to our LLVM
repo and CheckedC repo.  The Checked C tests will be kept in the CheckedC
repo so that we can use MIT license terms.  The repo can optionally be
added under projects\checkedc-llvm.  The tests will be be picked up
automatically from there.

The test checks that the new pointer types work as parameter types, local
variable types, reuturn types, and structure member types.  They also check
the case where the >> in ptr<ptr<T>> is tokenized as a right shift
operator.