Address CR feedback.
There was some concern in the code review about the way that the
code tested for checked array types.  This change addresses that
concern.  There are 4 different array types in the clang type
system:  incomplete arrays, constant-sized arrays, variably-sized
arrays, and dependent array types.  Only incomplete arrays and
constant-sized arrays can be checked, althought we may eventually
extend other arrays to be checked.

The code was testing the checked property only on constant-sized
arrays and variably-sized arrays.  For the other cases, the code
was falling into default cases or in some cases silently falling
out of switch statements.   To make the code clearer, to avoid
silent errors, and to future-proof it, the code now checks
all 4 array cases and handles them.  The code asserts in a few places
if a default case is reached unexpectedly for array types.

This change also clarifies that makeNestedArrayChecked does
not need to produce an error message for nested variably-szied
arrays and dependent-array types.  There will be an error
message produced for the outer type.

This update removes a few extraneous empty line deletions
or additions.