Only allow the Checked C extension flag for C. (#16)
This change only allows the -fcheckedc-extension flag to be used for C programs
in clang. The clang driver will reject the use of -fcheckedc-extension for
 other C family languages supported by clang, including C++, Objective
 C, OpenCL, and CUDA. This addresses issue #9 in the checked-clang Github repo.

We are currently not modifying clang to support these other languages, which
 is why need to disallow using the extension with them.

This change also updates the documentation on testing to reflect the
fact that there are now differences in the number of tests in the master and
baseline branches.  For testing changes, developers should make sure that no
new unexpected failures are introduced.  For updating to the latest sources
for LLVM/clang, developers should make sure that the additional Checked C
tests in the master branch pass.

Testing:
- Add 5 new tests to clang.  They test that use of the extension flag is
accepted for C files and rejected for C++, Objective C, CUDA, and 
OpenCL files..  This also applies to files that are compiled as a 
different file type than  their extension.
- Also test that clang-cl understands the Checked C flag when it is passed
in via -Xclang 
- As recommended by the clang documentation, I placed the tests in with other
similar tests in the test tree.