Parse bounds declarations for function return values.
This extends parsing of function declarators to parse bounds expressions for
function return values.  A return bounds expression is declared by following
a parameter list with ':' bounds-expression.

The return bounds information is propagated to the representation of function
declarations in clang.  It is not propagated yet to function types because
there is no way to represent it in function types.  I have opened issue #20
to track taht and added a few TODO's to the code code flagging where the
information will need to be added.

Testing:
- Add feature tests for parsing return bounds.  The tests are in in
the file return_bounds.c, which will be committed separately to the checkedc
Github repo.
- Passes existing clang regression test baseline.