Address code review feedback and improve an error message.
This change addresses code review feedback to rename some variables and add
a comment.

It also improves the error message for bounds(a,b) when a and b are omitted.
We would get two error messages at the same location, one saying that an
expression was expected and the other saying that a ',' was expected.  The
improvement is to print the error message for the ',' only if nothing
went wrong trying to parse the lower bound.