Parse bounds declarations for function return values. (#21)
This extends parsing of function declarators to parse bounds expressions for
function return values.  A return bounds expression is declared by following
a parameter list with ':' bounds-expression.

The return bounds information is propagated to the representation of function
declarations in clang.  It is not propagated yet to function types because
there is no way to represent it in function types.  I have opened issue #20
to track that and added a few TODO's to the code flagging where the
information will need to be added.

This change also improves the error message for bounds(a, b) when both
a and b are omitted.  We would generate two error messages, one saying that
an expression was expected and saying that ',' was expected.  The fix is to
generate the error message for ',' only if parsing of the lower bound 
expression (a) did not have an error.

Testing:
- Add feature tests for parsing return bounds.  The tests are in 
the file return_bounds.c, which will be committed separately to the checkedc
Github repo.
- Also added tests to return_bounds.c of bounds expressions with omitted
arguments. 
- Passes existing clang regression test baseline.