Fix crash during typo correction in C mode. (#22)
This fixes a crash in clang that can occur during typo correction in C mode.
I hit this crash while testing out parsing of incorrect bounds declarations on
structure members.

The crashes were fixed in the mainline clang sources for binary operators
with SVN r272587.  I've modified the fix to cover conditional operators as well.
The modification is out for code review for the mainline clang sources.  I'm
checking in the modified fix to the Checked C clang branch now.