Fix type tests for BoundsExpr and bounds-safe interface checks.
The type test implementation for BoundsExpr was broken.  There were two
problems:
* We were missing the static classof methods that the template
 implementation requires.
* The bit fields for CountBoundsExpr and NullaryBoundsExpr weren't
  padded, so were accidentally overwriting the kind field.

Also add checks to specific to bounds-safe interface.  Specifcally,
declaring bounds for local variables that are unchecked pointer types
or unchecked array types is not allowed.

Shorten some error messages so that they fit on an 80 charater line.