Fix test failures.
This change fixes a few tests failures seen when running check-all:

1. The Checked C rewriter tool failed to build:
- FunctionDecl params has been renamed to parameters().
- PrintStackTraceonErrorSignal now takes at least one parameter.
2. For Checked C, code for printing an error message was duplicated during
the merge.  Remove the duplicate code.
3. Two tests that the -fcheckedc-extension flag cannot used for a CUDA
file were failing because CUDA was not installed.  The fix was to add
flags to the RUN line to not link with the CUDA library or include CUDA
header files.

Testing: check-all succeeds with no unexpected test failures.