Revamp kinds for bounds expressions.
Instead of having separate kind enumerations for different bounds expression
classes, have just one kind enumeration for all bounds expression classes.
This makes it easier to write predicates that check for a specific kind of
bounds expression, without having to do a dynamic type test and cast to the
specific bounds expression type.  This addresses Github issue #50.

The downside of the unified kind enumeration is that a developer can use a
kind that is invalid for a specific bounds expression class.  Add code to
validate the kind information.

Add predicates for testing for specific kinds of bounds expressions and
clean up code for type checking bounds expressions by using the predicates.

This also adds a kind for the `: ptr` interop bounds annotation to the
kind enumeration for bounds expressions.

Testing:
- Passes existing clang and Checked C tests.