Change Checked C keywords to avoid conflicts with existing identifiers.
This change modifies the existing Checked C keywords to start with
an underscore and a capital letter.  This avoids conflicts with existing
C identifiers.  This addresses issue #53.

This change follows the rules in the C Standard: all identifiers beginning
with an underscore followed by a capital letter or underscore are reserved.
This allows them to be used for language extensions without conflicting with
existing identifiers. Of course, some programs do not follow the standard
and use these identifiers.