Change Checked C keywords to avoid conflicts with existing identifiers. (#54)
This change modifies the existing Checked C keywords to start with an underscore and a capital letter.  This avoids conflicts with existing C identifiers.  This addresses issue #53.

This change follows the rules in the C Standard: all identifiers beginning with an underscore followed by a capital letter or underscore are reserved.  This allows them to be used for language extensions without conflicting with existing identifiers. Of course, some programs do not follow the standard
and use these identifiers.

There is a corresponding change to the GitHub Checked C repo that adds a header "stdchecked.h" that defines macros that map the original keywords to the new standards-compliant keywords.

Testing:
- The Checked C repo tests will be updated separately to use stdchecked.h.
- Thee existing clang and Checked C tests pass.