Parse and represent bounds-safe interface type annotations (#55)
This change implements parsing of bounds-safe interface type annotations.  The type annotations are a generalization of the `int * x : ptr` annotation described in the specification.    That annotation does not provide a way to represent information for pointers with multiple levels of indirection (`int **x` for example).   Instead of just using `ptr`, we will allow a checked pointer type be specified.   The annotation for something like `int **` could be `_Ptr<_Ptr<int>>`.   This richer form is needed so that we can represent bounds-safe interfaces where pointers are passed by reference, for example.'

A bounds-safe type annotation has the syntax `type` `(` _type name_ `)` where _type name_
is a C type name.   _type name_ can include typedef'ed names and type modifies.  The type annotation
can be used in place of a bounds expression (following a `:`).   For example:
```
void f(int *p : type(ptr<int>)) {
}
```
We create a subclass of `BoundsExpr` to represent the new annotation, even though this annotation is not really a bounds expression.   It will be needed in the same places in the compiler where bounds expressions are needed.

The bounds-safe interface type annotations do not need to be delay parsed for parameters or structure/union members.   They don't refer to other parameters or members.  We special case this in the parser.   This simplifies the code because we don't have to implement delay parsing for types.

For structure/union members, we now store the bounds annotation or the tokens to be delayed parsed in the FieldDeclarator structure (which already stores the bitsize expression for a member).   This lets us revert changes to the callback method that passed the tokens as extra arguments.

This change also fixes a bug that I noticed in stmt serialization.  We weren't writing out the kind field for RangeBoundsExpr.   It is possible to have an invalid kind field (this could happen if someone serializes IR with invalid bounds expressions.  I also cleaned up serialization code that needs to access some private members.  We had setter methods - I noticed the rest of the serialization code is using friend classes instead.

Testing:
- Passes existing clang and Checked C tests.
- A new test for parsing these annotations will be added to the Checked C repo by a separate commit.