First cut of extending type checking to follow interop rules.
This works and passes  initial testing.  At variable assignments,
member assignments, and function calls, a source value with a checked pointer
type can be converted to an unchecked pointer type if a bounds-safe
interface annotation is present on the target.  Assigning the value to
target with the type specified or implied by the bounds-safe interface must be
legal following the C typechecking rules.