Implement implicit type conversions for bounds-safe interfaces (#62)
This change implements implicit conversions from checked pointers to unchecked pointers at bounds-safe interfaces.  These conversions may happen at assignments to variables or members and at function calls.

The conversions are mostly straightforward to implement.  The code needs to choose whether to use the original type or the interoperation type.  This change adds code to SemaExpr.cpp and SemaInit.cpp that makes this decision.    Because the code is quite similar in both cases, we do not put the decision-making inline around the current calls to CheckSingleAssignmentConstraints.  Instead we factor it into a function called ResolveSingleAssignmentType.  This also helps deal with fact that CheckSingleAssignmentContraints is tricky to use: it has default parameters that control whether it emits diagnostics and side-effects the right-hand side expression.   ResolveSingleAssignmentType needs to try each potential left-hand side by calling CheckSingleAssignmentConstraints.  It invokes CheckSingleAssignmentContraints with parameters that prevent side-effects. 

This change also adds code that computes the Checked C interop type for a variable, member, or expression being initialized.   Some care is needed when handling parameters with unchecked array types and bounds-safe interfaces.  clang retypes those parameters so that they have pointer types (the outermost array type is replaced with a pointer type).   If the checked type is being inferred from an unchecked type (because of the presence of a bounds expression),  the inference uses the original
unadjusted array type.  It converts that array type to its checked array equivalent.   For bounds-safe interface type annotations, the annotated type is expected to be a checked array type.  
The new array types are adjusted to match the retyping of array parameters using the getAdjustedParameterType method.

Testing:
- A new file interop.c will be committed separately to the Checked C repo.  The file has tests for implicit conversions at bounds-safe interfaces.  It will be added to the test\typechecking subdirectory.
- Passes existing clang and Checked C tests.
- A null pointer bug was found in the bounds-safe interface implementation while running the Checked C rewriter regression tests.   This was a bug specific to the clang implementation of  Checked C, so it does not  belong in the language conformance tests in the Checked C repo.  Created a new CheckedC  directory under clang\tests and placed a cut-down repro of the bug there.