Fix typos in methods recognizing starts of bounds expressions and type annotations. (#69)
This change fixes typos in methods recognizing the starts of bounds expressions and interop
type annotations.   This addresses issue #67.

The methods StartBoundsExpression and StartInteropTypeAnnotation both have a parameter named tok.  There is also a global variable named Tok that holds the current token.  The bodies of these functions are mistakenly using Tok instead of tok when getting the information for an identifier.  This
results in these methods misbehaving when passed in a token that is not the current token.

The fix is to rename the parameter to something other than tok so that we do not have variables whose names only differ in case.

Testing:
- Passed existing Checked C regression tests.