Better handling of syntax errors in return bounds declarations. (#70)
This change improves the error messages and recovery for syntax errors in
return bounds expressions.

Return bounds declarations are placed immediately after the argument list
in a function declarator, using the syntax `:` bounds-expression.  If the
bounds expression is syntactically incorrect, we would like to recover from
this error so that we can generate as many sensible error messages as
possible.   In particular, a function body might immediately follow
the bounds expression.

The code for parsing a function declarator tried to recover by scanning to
the body of a function body or a semi-colon.  This is not the right recovery
action. First, a function declarator can be embedded inside another declarator.
Second, this may be a semantic failure because clang does not differentiate
between semantic vs. parsing failures in the AST.  In that case, the code might
skip syntactically valid tokens, leading to confusing error messages.

This change moves the code for recovering from semantic/parsing failures for
return bounds expressions from function declarator parsing to
Parser::ParseDeclGroup.  This is the code in clang that recognizes whether
the parser is parsing a function body or a function declaration.  This is
a better place for error recovery because there is more context on what to
expect next.

We look for two cases: (1) the parsing/semantic analysis of the return bounds
expression failed and there is not a valid beginning of a function body or
end of function declaration (2) a misplaced bounds expression is encountered.
In the first case, the code tries to recover by doing the prior error action
of scanning to the beginning of the function body.   In the second case, we
guess that the user meant the bounds expression to be a return bounds
expression for the function declarator. The code emits an appropriate
diagnostic and points out where the bounds expression might be moved to.
The code then bails out on further processing of the function declaration or
body.

We do not add any code for recovery for the case where the function
declarator is nested within another declarator.  It is hard to know what
to do in this situation.  We could try to scan to the next token for the
enclosing declarator (a left parenthesis, a left bracket, or a right
parenthesis if the enclosing parameter is a paranethesized declarator.
It's not clear any of them put the parser in a sensible recovery position
in the case of a syntax error in a bounds expression, which could be any
syntactic form for a C expression. For now, we do nothing special here.

Testing:
* Passes existing clang tests.
* Passes existing Checked C tests for syntax errors in return bounds
  expressions.
* Add new tests to tests/parsing/return_bounds.c in the Checked C Github
  repo for a misplaced return bounds expression after a complex function
  declarator.  In a complex function declarator, the function declarator is
  nested within other declarators that construct the function return type.
  These tests will be committed separately.