Check bounds-safe interface type annotations.
This change adds the initial version of checking of bounds-safe interface
type annotations. For bounds-safe interface types, we require that:
 1. The bounds-safe interface type must be a checked pointer or array type.
 2. The declared type and the bounds-safe interface type be compatible
    types when the checked property is ignored on pointer types and array types.
 3. The bounds-safe interface type be at least as checked as the original
     declared type.

This change adds new methods for types to compute properties 2 and 3.
It adds code to the parsing of bounds-safe interop types to enforce
property 1.  The three properties imply that the declarations with
a bounds-safe interop type must have a pointer or array type too.

The error messages for bounds interop type annotations can be improved
when a declaration that does not have a pointer or array type is
a given a bounds-safe interface types.  The messages complain about
the bounds-safe interface type or about the types not being compatible,
when they should flag that the declaration cannot have a bounds-safe
interface.  That'll be a follow-up commit.

The code in lib/Parse/ParseDecl.cpp for setting an interop type annotation for
a structure/union member just set the bounds expression on the FieldDecl.
It did not invoke ActOnBoundsDecl, which caused it to miss new checking
 adding to ActOnBoundsDecl.  The code for other bounds expressions did invoke
ActOnBoundsDecl. (There are two separate code paths because bounds
interop types are easily parsed and bounds expressions are lazily parsed.)