Address code review feedback.
This change addresses code review feedback, making some parameters
const, fixing some white space, and improving a confusing comment.

I also noticed that the error handling in event of a syntax error or problem
with a type name was incorrect.  We should fall through to the method that
looks for a closing paren.  It provides better diagnostics and error
recovery than just skipping to a parentheses or semi-colon.