Allow type qualifiers and static keyword in parameter type annotations. (#73)
This change extends the parsing of interop type annotations for parameters.  The code now allows the static keyword or type qualifiers to occur inside the first dimension of an array type in an interop type annotation, just as they can occur for regular parameter types.

This required adding the declarator as an argument to some parsing methods for interop type annotations, so that the code can tell when it is in the special context of parsing an interop type   annotation for a parameter declaration.

The code for parsing type names allows a Context value to be passed in.  It can fail an assert in SemaType.cpp if a Context other than TypeName is passed in, though.  Other contexts can allow variable names to occur within the parsing of  a type name.  Modify the assert in SemaType.cpp to apply to only the TypeName context and issue an error message if an identifier is used in other contexts.  I weighed adding a new context for parsing interop bounds expression, but that change seem worse, requiring fixing up a bunch of places in the code.  It seemed better to generalize parsing of type names slightly instead.

This fixes a few places where existing code for Checked C exceeded the 80 character line limit for the LLVM coding convention.  The parameter IsReturn is intentionally unused.  This change was extracted from a larger commit that uses it.

Testing:
- Passes clang regression tests.
- Passes existing Checked C regression tests.
- Added new tests for parsing interop bounds declarations with array types to the Checked C GitHub repo.  The tests will be added to tests/parsing/interop_types.c.  They will be committed separately.