Check bounds-safe interface type annotations. (#74)
This change adds checking of bounds-safe interface type annotations.  It addresses issue #64.

The type in the bounds-safe interface type annotation is called the bounds-safe interface type.  For a bounds-safe interface type annotation, we require that:
1. The variable to which the annotation is attached have a pointer or array type.  For annotations on return values, the return type must be a pointer type.
2. The  bounds-safe interface type be a pointer or array type.
3. The bounds-safe interface type be a checked type.  A checked type defined inductively.  A checked type is (a) a checked pointer or array type, (b) a pointer type whose referent type is a checked type, or (c) an array type whose element type is a referent type, or (d) a function type whose return type is a checked type or that has a parameter whose type is a checked type.
4.  The declared type and the bounds-safe interface type be compatible types when the checked property is ignored on pointer types and array types.
5.  The bounds-safe interface type be at least as checked as the original  declared type.

This change adds new methods for types to compute properties 4 and 5.   There is already a method for computing property 3.  The code for checking bounds annotation is added at the point where a bounds-safe interface type is attached to a declaration.   

It took a bit of effort to come up with accurate and helpful error messages. I decided to have more general messages that gloss over some of the subtle points.  For example, for property 4, the error message says that there is a mismatch between the interface and declaration type, and doesn't add the subtlety that it is compatibility modulo checkedness.  For property 3, the error message says that the type must be checked.    The message points at the problematic interface type.  For simple cases, such as omitting the checked modifier in an array type or not using a checked type, this will help the user.  We actually allow types that are mix of checked and unchecked pointers, but that subtlety is omitted.  When there are nested levels of pointers, we don't know which pointers are intended to be checked.  Also, declarations may have multiple problems.  An interface type might not be a checked type and it might not be compatible.    

The code makes a best effort to emit multiple diagnostics if there are independent errors.   This will avoid an unnecessary round of recompilation in some cases. The code does not emit errors that are implied by other errors.  For example, if the interface type is an illegal type (not a pointer or an array), there's no point in pointing in a message that says that declared type and the interface type are incompatible. 

The code in lib/Parse/ParseDecl.cpp for setting an interop type annotation for a structure/union member just set the bounds expression on the FieldDecl.  It did not invoke ActOnBoundsDecl, which caused it to miss new checking adding to ActOnBoundsDecl.  The code for other bounds expressions did invoke ActOnBoundsDecl. (There are two separate code paths because bounds interop types are easily parsed and bounds expressions are lazily parsed.)

Testing:
- An extensive set of new feature tests will be added via a separate commit to the Checked C github repo.
- Added a few clang tests that check that multiple error messages are emitted for independent problems in a bounds-safe interface type declaration.
- Passed existing Checked C tests.
- Passed existing clang regression tests.