Checked C type compatibility rules for no prototype functions.
This changes implements Checked C type compatibility rules for no prototype
functions. Calls to functions without prototypes are not typechecked, so we
restrict the usage of checked types with no-prototype functions.

A function type without a prototype is not compatible with a
function type with a prototype if the prototype takes a checked argument.
The Checked C specification defines this in more depth.  In addition,
a function type without a prototype cannot return a checked value.

This change alters the implementation of type compatibility to
enforce this rules.  When two declarations of a function are
incompatible, we try to diagnose whether the failure is due to a Checked C
rule and print descriptive error messages.  When the failure is simply
becaues of the presence or absence of a prototype of a declaration,
the messages work well.  When the failure is due to a mismatch between
parameters with function pointer types, we don't try to explain what happened
at this point.

The change also updates the construction of function types from function
declarators to enforce the rules about return types.

clang works by seeing if types are compatible and then merging
the types so that the latest declaration has the merged type.  For the
Checked C rules, this breaks down in one case: a function can be declared
as no prototype and then be declared to have a prototype that uses an
 incomplete structure type.   Those are compatible under the Checked C rules
because the function is incomplete.  The type can then be completed and become
a checked type (because it has a checked member).

Instead of trying to revisit function types whenever types are completed,
we take advantage of an aspect of C that no function can be defined with
incomplete types for parameters or return types.  When we see a definition,
we look back all the prior declarations of the function to see if any of them
were a no prototype declarations.

The code misses the case of no-prototype function pointers being
incompatible with prototype function pointers that take integer
arguments with bounds.  This is because of issue #20 (adding bounds to
function types). We will fix that as part of addressing  issue #20.
The code will also become simpler after issue #20 is completed.

Testing:
- This change uncovered places in existing tests in the Checked C repo
  where we were using no prototype functions.  Update the tests to properly
  declare the functions as having no arguments (using the f(void) ... pattern).
- Add a new tests on to Checked C repo in typechecking\no_prototype_functions.c.
- Add some clang-specific tests under tests\CheckedC\typechecking.c.
- Passes the updated and new Checked C tests.
- Passes the existing clang regression tests.